Introduce Init method for TxServiceModule to replace copy. (#104)

+++ src/cc/local_cc_shards.cpp
@@ -4499,7 +4499,8 @@ void LocalCcShards::DataSyncForHashPartition(
                 {
                     // cce_ is null means the key is already persisted on kv, so
                     // we don't need to put it into the flush vec.
-                    int32_t part_id = (rec.Key().Hash() >> 10) & 0x3FF;
+                    int32_t part_id =
+                        static_cast<int32_t>((rec.Key().Hash() >> 10) & 0x3FF);
                     if (table_name.Engine() == TableEngine::EloqKv)
                     {
                         data_sync_vec->emplace_back(
